refactor: rename lib/cli/cli-errors.jsto lib/cli/cli-validation-errors.js
